functionality of vote system was created

+++ Web Pages/appchallengers/src/main/java/com/appchallengers/webservice/dao/dao/UserDao.java
@@ -7,10 +7,10 @@ public interface UserDao {
     Users saveUser(Users user);
     Users findByEmail(String email);
     Long checkEmail(String email);
-    Long checkIdAndPasswordSalt(Integer id,String passwordSalt);
-    void confirmEmail(Integer id,String passwordSalt);
-    void changePassword(Integer id,String passwordHash);
-    Users findUserById(Integer id);
+    Long checkIdAndPasswordSalt(long id,String passwordSalt);
+    void confirmEmail(long id,String passwordSalt);
+    void changePassword(long id,String passwordHash);
+    Users findUserById(long id);
     Long login(String email,String passwordHash);
 
 }
