+++ Web Pages/appchallengers/src/main/java/com/appchallengers/webservice/model/response/RelationshipResponse.java
@@ -4,10 +4,10 @@ public class RelationshipResponse {
 
     private Integer firstUserId;
     private Integer secondUserId;
-    private Integer actionUserId;
+    private long actionUserId;
     private Integer relationshipStatus;
 
-    public RelationshipResponse( Integer firstUserId, Integer secondUserId, Integer actionUserId, Integer relationshipStatus) {
+    public RelationshipResponse(Integer firstUserId, Integer secondUserId, long actionUserId, Integer relationshipStatus) {
         this.firstUserId = firstUserId;
         this.secondUserId = secondUserId;
         this.actionUserId = actionUserId;
